Move SImplexIndex to vr subpackage
