+++ src/main/scala/io/github/jakipatryk/sparkpersistenthomology/internal/vr/SimplexIndex.scala
@@ -1,4 +1,4 @@
-package io.github.jakipatryk.sparkpersistenthomology.internal.utils
+package io.github.jakipatryk.sparkpersistenthomology.internal.vr
 
 /** Represents a unique index of a simplex in the filtration.
   *
