+++ src/test/scala/io/github/jakipatryk/sparkpersistenthomology/internal/vr/CliqueIteratorSpec.scala
@@ -1,9 +1,6 @@
 package io.github.jakipatryk.sparkpersistenthomology.internal.vr
 
-import io.github.jakipatryk.sparkpersistenthomology.internal.utils.{
-  CombinatorialNumberSystem,
-  SimplexIndex
-}
+import io.github.jakipatryk.sparkpersistenthomology.internal.utils.CombinatorialNumberSystem
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
